fix: cast the two enum values the inspector passes as i32
The same trap as the seek constants: libvlc's enums are typed by bindgen per target, and the
inspector passes them into this binding's own \#[func]\, whose parameters are \i32\. Windows
was happy because \c_int\ is \i32\ there; Linux and Android were not. The cast, and the
allowance that goes with it, are what the rest of the crate already does for this.

+++ src/vlc_media_inspector.rs
@@ -115,6 +115,7 @@ impl IEditorProperty for VlcMediaInspector {
     }
 }
 
+#[allow(clippy::unnecessary_cast)]
 impl VlcMediaInspector {
     /// Lets go of the media this control was showing, so that it stops reporting into a control
     /// that has moved on to another one.
@@ -203,7 +204,7 @@ impl VlcMediaInspector {
             THUMBNAIL_SIZE,
             THUMBNAIL_SIZE,
             false,
-            libvlc_picture_type_t_libvlc_picture_Png,
+            libvlc_picture_type_t_libvlc_picture_Png as i32,
             5000,
         );
         if self.request.is_none() {
